+++ pkg/simulation/manager/manager.go
@@ -9,7 +9,7 @@ import (
 )
 
 // Creates default user array of the specified size. Panics if there is not enough containers or the nextfunc is nil.
-func MakeDefaultSimulation(count int, clientContainers []*container.Container, nextfunc func(*Behavior.SimpleHumanTraits) float64) []*User.SimulatedUser {
+func MakeDefaultSimulation(count int, clientContainers []*container.Container, nextfunc func(*Behavior.SimpleHumanTraits) int) []*User.SimulatedUser {
 	if len(clientContainers) < count {
 		panic(fmt.Sprintf("Insufficient number of clientContainers provided as argument. Expected %v, got %v", count, len(clientContainers)))
 	}
@@ -29,7 +29,7 @@ func MakeDefaultSimulation(count int, clientContainers []*container.Container, n
 }
 
 // Uses the supplied options struct to generate users. If any critical option is nil, the function will return default users. Panics if there is not enough containers or the nextfunc is nil.
-func MakeSimUsersFromOptions(count int, clientContainers []*container.Container, nextfunc func(*Behavior.SimpleHumanTraits) float64, options *Types.SimUserOptions) []*User.SimulatedUser {
+func MakeSimUsersFromOptions(count int, clientContainers []*container.Container, nextfunc func(*Behavior.SimpleHumanTraits) int, options *Types.SimUserOptions) []*User.SimulatedUser {
 	if len(clientContainers) < count {
 		panic(fmt.Sprintf("Insufficient number of clientContainers provided as argument. Expected %v, got %v", count, len(clientContainers)))
 	}
@@ -65,7 +65,7 @@ func MakeSimUsersFromOptions(count int, clientContainers []*container.Container,
 }
 
 // Alice, Bob, Charlie and Dorothy example only sending regular messages. Panics if there is not enough containers or the nextfunc is nil.
-func MakeAliceBobRegularExampleSimulation(clientContainers []*container.Container, nextfunc func(*Behavior.SimpleHumanTraits) float64) []*User.SimulatedUser {
+func MakeAliceBobRegularExampleSimulation(clientContainers []*container.Container, nextfunc func(*Behavior.SimpleHumanTraits) int) []*User.SimulatedUser {
 	if len(clientContainers) < 4 {
 		panic(fmt.Sprintf("Insufficient number of clientContainers provided as argument. Expected %v, got %v", 4, len(clientContainers)))
 	}
@@ -86,7 +86,7 @@ func MakeAliceBobRegularExampleSimulation(clientContainers []*container.Containe
 }
 
 // Alice, Bob, Charlie and Dorothy example sending both regular messages and deniable messages with bursting. Panics if there is not enough containers or the nextfunc is nil.
-func MakeAliceBobDeniableBurstExampleSimulation(clientContainers []*container.Container, nextfunc func(*Behavior.SimpleHumanTraits) float64) []*User.SimulatedUser {
+func MakeAliceBobDeniableBurstExampleSimulation(clientContainers []*container.Container, nextfunc func(*Behavior.SimpleHumanTraits) int) []*User.SimulatedUser {
 	if len(clientContainers) < 4 {
 		panic(fmt.Sprintf("Insufficient number of clientContainers provided as argument. Expected %v, got %v", 4, len(clientContainers)))
 	}
@@ -104,7 +104,7 @@ func MakeAliceBobDeniableBurstExampleSimulation(clientContainers []*container.Co
 }
 
 // Alice, Bob, Charlie and Dorothy example sending both regular messages and deniable messages, but without bursting. Panics if there is not enough containers or the nextfunc is nil.
-func MakeAliceBobDeniableNoBurstExampleSimulation(clientContainers []*container.Container, nextfunc func(*Behavior.SimpleHumanTraits) float64) []*User.SimulatedUser {
+func MakeAliceBobDeniableNoBurstExampleSimulation(clientContainers []*container.Container, nextfunc func(*Behavior.SimpleHumanTraits) int) []*User.SimulatedUser {
 	if len(clientContainers) < 4 {
 		panic(fmt.Sprintf("Insufficient number of clientContainers provided as argument. Expected %v, got %v", 4, len(clientContainers)))
 	}
